- Implemented the speed optimisations in a slightly more conservative way due to problems with some benchmarks - Changed IPK cache size to 256 entries - Changed instrumentation code to use 64-bit event counters

+++ src/Utility/Instrumentation/BackgroundMemory.cpp
@@ -8,10 +8,10 @@
 #include "../Parameters.h"
 #include "BackgroundMemory.h"
 
-int BackgroundMemory::numReads_ = 0;
-int BackgroundMemory::numWrites_ = 0;
-int BackgroundMemory::numWordsRead_ = 0;
-int BackgroundMemory::numWordsWritten_ = 0;
+unsigned long long BackgroundMemory::numReads_ = 0;
+unsigned long long BackgroundMemory::numWrites_ = 0;
+unsigned long long BackgroundMemory::numWordsRead_ = 0;
+unsigned long long BackgroundMemory::numWordsWritten_ = 0;
 
 void BackgroundMemory::read(MemoryAddr address, uint32_t count) {
 	numReads_++;
@@ -23,10 +23,10 @@ void BackgroundMemory::write(MemoryAddr address, uint32_t count) {
 	numWordsWritten_ += count;
 }
 
-int BackgroundMemory::numReads()			{return numReads_;}
-int BackgroundMemory::numWrites()			{return numWrites_;}
-int BackgroundMemory::numWordsRead()		{return numWordsRead_;}
-int BackgroundMemory::numWordsWritten()		{return numWordsWritten_;}
+unsigned long long BackgroundMemory::numReads()			{return numReads_;}
+unsigned long long BackgroundMemory::numWrites()			{return numWrites_;}
+unsigned long long BackgroundMemory::numWordsRead()		{return numWordsRead_;}
+unsigned long long BackgroundMemory::numWordsWritten()		{return numWordsWritten_;}
 
 void BackgroundMemory::printStats() {
 	if (BATCH_MODE) {
@@ -37,8 +37,8 @@ void BackgroundMemory::printStats() {
 	}
 
 	if (numReads_ > 0 || numWrites_ > 0) {
-		int accesses = numReads_ + numWrites_;
-		int words = numWordsRead_ + numWordsWritten_;
+		unsigned long long accesses = numReads_ + numWrites_;
+		unsigned long long words = numWordsRead_ + numWordsWritten_;
 
 		cout <<
 			"On-chip scratchpad:\n" <<
