+++ src/Utility/Instrumentation/IPKCache.h
@@ -22,17 +22,17 @@ public:
   static void read(const ComponentID& core);
   static void write(const ComponentID& core);
 
-  static int  numTagChecks();
-  static int  numHits();
-  static int  numMisses();
-  static int  numReads();
-  static int  numWrites();
+  static unsigned long long  numTagChecks();
+  static unsigned long long  numHits();
+  static unsigned long long  numMisses();
+  static unsigned long long  numReads();
+  static unsigned long long  numWrites();
 
   static void printStats();
 
 private:
 
-  static int numHits_, numMisses_, numReads_, numWrites_;
+  static unsigned long long numHits_, numMisses_, numReads_, numWrites_;
 
 };
