+++ src/Utility/Instrumentation/Registers.cpp
@@ -8,10 +8,10 @@
 #include "../../Datatype/ComponentID.h"
 #include "Registers.h"
 
-int Registers::numReads_ = 0;
-int Registers::numWrites_ = 0;
-int Registers::numForwards_ = 0;
-int Registers::stallRegs_ = 0;
+unsigned long long Registers::numReads_ = 0;
+unsigned long long Registers::numWrites_ = 0;
+unsigned long long Registers::numForwards_ = 0;
+unsigned long long Registers::stallRegs_ = 0;
 
 // TODO: determine which reads/writes were wasted, and could have been replaced
 // by explicit data forwarding.
@@ -23,10 +23,10 @@ void Registers::write(const ComponentID& core, RegisterIndex reg)   {numWrites_+
 void Registers::forward(const ComponentID& core, RegisterIndex reg) {numForwards_++;}
 void Registers::stallReg(const ComponentID& core)                   {stallRegs_++;}
 
-int  Registers::numReads()      {return numReads_;}
-int  Registers::numWrites()     {return numWrites_;}
-int  Registers::numForwards()   {return numForwards_;}
-int  Registers::stallRegUses()  {return stallRegs_;}
+unsigned long long  Registers::numReads()      {return numReads_;}
+unsigned long long  Registers::numWrites()     {return numWrites_;}
+unsigned long long  Registers::numForwards()   {return numForwards_;}
+unsigned long long  Registers::stallRegUses()  {return stallRegs_;}
 
 void Registers::printStats() {
   if (BATCH_MODE) {
