+++ src/Utility/Instrumentation/Stalls.h
@@ -19,16 +19,16 @@ class Stalls: public InstrumentationBase {
 
 public:
 
-  static void stall(const ComponentID& id, int cycle, int reason);
-  static void unstall(const ComponentID& id, int cycle);
-  static void idle(const ComponentID& id, int cycle);
-  static void active(const ComponentID& id, int cycle);
+  static void stall(const ComponentID& id, unsigned long long cycle, int reason);
+  static void unstall(const ComponentID& id, unsigned long long cycle);
+  static void idle(const ComponentID& id, unsigned long long cycle);
+  static void active(const ComponentID& id, unsigned long long cycle);
   static void endExecution();
 
-  static int  cyclesActive(const ComponentID& core);
-  static int  cyclesIdle(const ComponentID& core);
-  static int  cyclesStalled(const ComponentID& core);
-  static int  executionTime();
+  static unsigned long long  cyclesActive(const ComponentID& core);
+  static unsigned long long  cyclesIdle(const ComponentID& core);
+  static unsigned long long  cyclesStalled(const ComponentID& core);
+  static unsigned long long  executionTime();
 
   static void printStats();
 
@@ -42,7 +42,7 @@ public:
 private:
 
   // The times that each cluster started stalling.
-  static std::map<ComponentID, int> startedStalling;
+  static std::map<ComponentID, unsigned long long> startedStalling;
 
   // The reason for each cluster being stalled at the moment (if stalled).
   static std::map<ComponentID, int> stallReason;
@@ -55,7 +55,7 @@ private:
 
   // The times that each cluster became idle (or UNSTALLED if the cluster is
   // currently active).
-  static std::map<ComponentID, int> startedIdle;
+  static std::map<ComponentID, unsigned long long> startedIdle;
 
   // The total number of cycles each cluster has spent idle.
   static CounterMap<ComponentID> idleTimes;
